# openapi.yaml
info:
    description: 查询凡高和文森特项目的开票、回款、成本、利润情况
    title: 项目信息查询服务
    version: v2
openapi: 3.0.1
paths:
    /auth/scope:
        get:
            operationId: getAuthScope
            parameters:
                - in: query
                  name: memberId
                  required: false
                  schema:
                      type: string
                  description: 人员ID，跨域时 cookie 无法携带时通过此参数传递，与 cookie 中的 avatarImageUrl 值相同
            responses:
                "200":
                    content:
                        application/json:
                            schema:
                                properties:
                                    companyScope:
                                        description: 可查看的公司范围，0=全部，1=仅凡高，2=仅文森特
                                        type: integer
                                        enum: [0, 1, 2]
                                type: object
                    description: 查询成功
                "401":
                    content:
                        application/json:
                            schema:
                                properties:
                                    error:
                                        type: string
                                type: object
                    description: 未获取到人员身份
                "403":
                    content:
                        application/json:
                            schema:
                                properties:
                                    error:
                                        type: string
                                type: object
                    description: 无权限查看任何公司项目
            summary: 查询当前人员可查看的公司范围
    /intent:
        post:
            operationId: recognizeIntent
            requestBody:
                content:
                    application/json:
                        schema:
                            properties:
                                text:
                                    description: 用户输入的自然语言，如"大连紫悦府精装项目开票情况怎么样？"
                                    type: string
                            required:
                                - text
                            type: object
            responses:
                "200":
                    content:
                        application/json:
                            schema:
                                properties:
                                    projectKeyword:
                                        description: 提取出的项目关键词
                                        type: string
                                    queryType:
                                        description: 查询类型，invoice=开票，payment=回款，cost=成本，profit=利润，unknown=无法识别
                                        type: string
                                        enum: [invoice, payment, cost, profit, unknown]
                                type: object
                    description: 识别成功
                default:
                    description: ""
            summary: 从自然语言中识别项目关键词和查询类型
    /query/search:
        post:
            operationId: searchAndQuery
            requestBody:
                content:
                    application/json:
                        schema:
                            properties:
                                keyword:
                                    description: 项目关键词，如大连紫悦府
                                    type: string
                                queryType:
                                    description: 查询类型，必须传英文值，invoice=开票，payment=回款，cost=成本，profit=利润
                                    type: string
                                companyScope:
                                    description: 公司范围，0=全部（默认），1=仅凡高，2=仅文森特，从 /auth/scope 接口获取
                                    type: integer
                                    enum: [0, 1, 2]
                            type: object
            responses:
                "200":
                    content:
                        application/json:
                            schema:
                                properties:
                                    type:
                                        description: 结果类型，single=唯一匹配，multiple=多个匹配，not_found=未找到
                                        type: string
                                    result:
                                        description: 格式化后的查询结果文本，type=single时返回
                                        type: string
                                    message:
                                        description: 提示信息，type=not_found时返回
                                        type: string
                                    project:
                                        description: 匹配到的项目信息，type=single时返回
                                        properties:
                                            id:
                                                type: string
                                            name:
                                                type: string
                                            fullName:
                                                type: string
                                            company:
                                                type: string
                                        type: object
                                    projects:
                                        description: 匹配到的项目列表，type=multiple时返回
                                        items:
                                            properties:
                                                id:
                                                    type: string
                                                name:
                                                    type: string
                                                fullName:
                                                    type: string
                                                company:
                                                    type: string
                                            type: object
                                        type: array
                                type: object
                    description: 查询成功
                default:
                    description: ""
            summary: 搜索项目并查询信息
    /query/detail:
        post:
            operationId: queryByProjectId
            requestBody:
                content:
                    application/json:
                        schema:
                            properties:
                                projectId:
                                    description: 项目ID，从搜索接口获取
                                    type: string
                                company:
                                    description: 所属公司，凡高或文森特
                                    type: string
                                queryType:
                                    description: 查询类型，必须传英文值，invoice=开票，payment=回款，cost=成本，profit=利润
                                    type: string
                            type: object
            responses:
                "200":
                    content:
                        application/json:
                            schema:
                                properties:
                                    result:
                                        description: 格式化后的查询结果文本
                                        type: string
                                type: object
                    description: 查询成功
                default:
                    description: ""
            summary: 按项目ID精确查询
servers:
    - url: https://sd6sfgfrafnohot0pt020.apigateway-cn-beijing.volceapi.com